requirements.txt
baycomp==1.0.2
build==0.9.0
dataclasses==0.6
eventfd==0.2
face-recognition==1.3.0
future==0.18.2
keras-nightly==2.5.0.dev2021032900
lapsolver==1.1.0
libtmux==0.15.10
munch==2.5.0
opencv-python==4.6.0.66
patool==1.12
pycryptodome==3.15.0
tensorflow-privacy==0.8.7
torchac==0.9.3
torchdata==0.5.0
torchmetrics==0.10.3
torchtext==0.14.0
torchvision==0.14.0
